+++ CHANGELOG.md
@@ -132,3 +132,4 @@
 - 2023-10-14: Pool session, gp, scheduler, qualifier, +2
 - 2023-10-15: Metrics session, deals, qualifier, gp, +3
 - 2023-10-15: Err gp, qualifier, session, deals, +2
+- 2023-10-16: Builder qualifier, session, deals, gp, +3